new position estimation algorithm based on Expectation-Maximization
New position estimation algorithm based on Expectation-Maximization, written in Rust.
It's more robust than the previous algorithm and always runs in 3D. It guesses any
unknown coordinates in the given measurements (some of Apple's data doesn't contain
Z), resulting in better accuracy. It goes through RANSAC for each combo of 4 for a
maximum of 10 measurements, while any additional measurements will still be
candidates to be included in the inliers, which are the measurements used for
computing the final result.

On the Pixel 7a, the maximum time to estimate the position should be well under
100 ms under very conservative maximum load. In practice, it usually takes under
50 ms, while the old algorithm takes around 200 ms to get a location. When giving it
garbage data (real measurements but with X and Y being multiplied by 10,000), it
can even take up to 5 seconds before failing.

With the same garbage data, the new algorithm takes the same amount of time as it
does with real data (under 50 ms). Since early stopping is not implemented, this
represents the maximum possible amount of time.

This commit also adds the foundation for Wi-Fi RTT support for distance estimation.
As it is a work-in-progress, it isn't used for distance estimation yet. Wi-Fi
RTT should allow better accuracy than RSSI-based distance estimation.

Signed-off-by: Dmitrii <bankersenator@gmail.com>